hrApp/src/main/webapp/app/entities/job/route/job-routing-resolve.service.ts
import { Injectable } from '@angular/core';
import { HttpResponse } from '@angular/common/http';
import { Resolve, ActivatedRouteSnapshot, Router } from '@angular/router';
import { Observable, of, EMPTY } from 'rxjs';
import { mergeMap } from 'rxjs/operators';

import { IJob, Job } from '../job.model';
import { JobService } from '../service/job.service';

@Injectable({ providedIn: 'root' })
export class JobRoutingResolveService implements Resolve<IJob> {
  constructor(protected service: JobService, protected router: Router) {}

  resolve(route: ActivatedRouteSnapshot): Observable<IJob> | Observable<never> {
    const id = route.params['id'];
    if (id) {
      return this.service.find(id).pipe(
        mergeMap((job: HttpResponse<Job>) => {
          if (job.body) {
            return of(job.body);
          } else {
            this.router.navigate(['404']);
            return EMPTY;
          }
        })
      );
    }
    return of(new Job());
  }
}
